Refactor: Pull RuleTestBase and subclasses into separate file
... per feedback on a previous PR.
